Changed the way the error message related to unsupported currency is displayed
The new logic displays the error message as an admin notice, like it's done when client_id and client_secret are missing. This allows to inform the user that the plugin won't work, without preventing him from configuring it. Note: the gateway will still be unavailable for payment, as it was done before.